requirements.txt
openai
langchain
pdfminer.six
python-docx
streamlit
python-dotenv
